Add and integrate internal Device + Release trait
To begin generalizing over both SPI and I2C devices, we provide a
trait called Device. Device provides the same interface as the
MPU's internal methods read, read_many, write, and modify. It adapts
a SPI / nCS pair to provide the same implementation. We then
integrate this into the MPU definition.

We also define an initial I2CDevice that can adapt any I2C type
and implements the Device interface. This is just a prototype as
of this commit; we haven't provided an MPU interface that accepts
an I2C type.

At this commit, the RPi example should produce the exact same
outputs. The changes should be internal-facing only. I've no
ability to run the RPi example myself at this point, as I don't
have the hardware. I do have a BBBlue, which I'll be using to
test the I2C interface.

We're required to expose the Device, Releasable, and helper structs,
since they're part of the MPU public interface. As a result, we're
also required to expose the Register enumerations. I've hid them from
the documentation, since they should not be used by library users
but are required to be public.