Separate device- and sensor-specific functions
In this commit, we separate the MPU functions by device- and sensor-
specific functions. Consider a matrix resembling

| Device     | IMU | MARG | DON'T CARE |
| ---------- | --- | ---- | ---------- |
| SPI        |     |      |            |
| I2C        |     |      |            |
| DON'T CARE |     |      |            |

Our ideal is to put as much of the functionality into DON'T CARE,
DON'T CARE. When necessary, we specialize for IMU-only or MARG-only.
Behavior that is very much device- and sensor-specific is isolated
as such. This commit goes through that effort to begin isolating and
exposing the behaviors. There's no I2C device, yet...

This commit leaves the public API in a consistent state. Again,
the RPi example should build and run as before.